Strip invalid code points from the XML
I encountered XML with invalid code points in the wild. It is unable to
be decoded by the xml package. To work around this, I've decided to
strip such code points out, and then proceed with decoding.

This works by decoding from whatever charset we have to UTF-8, stripping
the code points, and then updating the XML header.

After doing this we can simplify the XML unmarshalling as we can say it
is always UTF-8 by that point.